requirements.txt
requests==2.32.*
environs==14.1.*
python-telegram-bot==13.2
urllib3==1.26.17
standard-imghdr==3.13.0
